fix(Loader): deprecate component (#5844)
This component is a remnant of temporary design guidelines and is not
recommended to be used anymore. The `BusyIndicator` should be the only
loading indicator available. Therefore, we will remove this component in
favor of the `BusyIndicator` with our next release.

For further information see [Fiori for Web Design
Guidelines](https://experience.sap.com/fiori-design-web/busy-handling/)