fix(AnalyticalTable): offer limited support for `infiniteScroll` combined with a grouped table (#6431)
Using `infiniteScroll` with a grouped table can result in an odd user
experience. For example, when rows are lazy-loaded via scrolling, new
rows might only be added to existing groups. This won't change the
scroll container's height, potentially forcing the user to scroll down
again to load more rows.
Therefore (and as we don't have a concept for this behavior) __it's not
recommended using a grouped table in combination with
`infiniteScroll`__!